fix hanging on ping error
If run with `-x` or anything not a valid URL, `ping` command (iputils) exits with an error code, but the generator is still trying to read more lines when there is none.

It (EOF) can also be checked with

    if not line:

But using `ping.poll()` is a safer option, because we want to know if there is an error, and if so, gping quits without error, although it should.