pinger: Canvas.horizontal_line should fill the remainder of line with space
Currently, if you call canvas.horizontal_line with a 'data' argument that is
shorter than the space to fill, you get a StopIteration exception.

The next function takes an optional second argument that is the default value
to return after iteration termiantes. With this change, we pass " " as the
default.