Refactor: Reduce iterator cloning in y_axis_bounds
By leveraging the `minmax` method provided by `itertools`, we were
able to eliminate the need for one iterator clone operation in the
`y_axis_bounds` function. This refactor enhances the efficiency and
readability of the codebase.

In #323, I made some mistakes in the `NoElements` branch and minmax
won't work on f64::NAN. This commit fixes the issue.

ref https://github.com/orf/gping/pull/323